+++ packages/wildness_ui/lib/wildness/wildness_properties.dart
@@ -42,32 +42,6 @@ final class WildnessProperties {
         maxScaleFactor: maxScaleFactor,
       );
 
-  /// Creates a [WildnessProperties] from [Components] instances for light and dark modes.
-  factory fromComponents({
-    Brightness? forceThemeMode,
-    Components? light,
-    Components? dark,
-    Configuration resources = const Configuration(),
-    Resources? lightResources,
-    Resources? darkResources,
-    ScrollPhysics physics = const ClampingScrollPhysics(),
-    double minScaleFactor = 0.5,
-    double maxScaleFactor = 1.2,
-  }) =>
-      WildnessProperties(
-        forceThemeMode: forceThemeMode,
-        components: Configuration.fromComponents(light: light, dark: dark),
-        resources: lightResources != null || darkResources != null
-            ? Configuration.fromResources(
-                light: lightResources,
-                dark: darkResources,
-              )
-            : resources,
-        physics: physics,
-        minScaleFactor: minScaleFactor,
-        maxScaleFactor: maxScaleFactor,
-      );
-
   /// Components of this theme.
   ///
   /// To define components, pass an [Iterable] containing one or more [WildnessBase] and their kinds